M2/requirements.txt
Flask
PyYAML
Werkzeug 